% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normic_colors.R
\docType{data}
\name{normic_colors}
\alias{normic_colors}
\title{Nordic Microbes A/S official colors}
\format{
An object of class \code{list} of length 9.
}
\usage{
normic_colors
}
\description{
A list of continuous and discrete color vectors created from "#609947", "#89b03c", "#c0d441", "#e8a226", "#efc42e", "#725b99", "#9b69aa"
}
\keyword{datasets}
